requirements.txt
django==4.1.7
djangorestframework==3.14.0
python-dotenv==0.21.1
psycopg2-binary==2.9.5
stripe==5.2.0
